cleanup so the app compiles, but TIDE is not yet implemented in other features as of yet

+++ app/src/main/kotlin/com/example/maiplan/network/api/CategoryApi.kt
@@ -47,7 +47,7 @@ interface CategoryApi {
     suspend fun createCategory(@Body categoryCreate: CategoryCreate): Response<Unit>
 
     @GET("categories/get-all-category")
-    suspend fun getAllCategories(@Query("user_local_id") userLocalId: Int): Response<List<CategoryResponse>>
+    suspend fun getAllCategories(@Query("user_local_id") userLocalId: Long): Response<List<CategoryResponse>>
 
     @POST("categories/update-category")
     suspend fun updateCategory(@Body category: CategoryResponse): Response<Unit>
